rename package machine/usb2 to machine/usb
