+++ src/machine/usb/dcd.go
@@ -1,4 +1,4 @@
-package usb2
+package usb
 
 import "unsafe"
