+++ src/machine/usb/hcd.go
@@ -1,4 +1,4 @@
-package usb2
+package usb
 
 type hcd interface {
 	class() class
